17/04 Update
Changed manipulation of case to meet requirements for assessment,
Additional function for performing rotational decryption based on statistical analysis (needs to be modified so that the cipher is valid in even cases)
began initial work on function that will perform analysis of string encrypted with substitution cipher, compare the frequency of the letters used with general letter use in english language and generate an initial key to match. This key will then evaluate itself by comparing a string created using the key with a dictionary of common words.