OnlyCommentsAndTestingRequired
Reverted back to unsubdecryption writing to file as the functions allowing decryption without writing to file didn't provide any speed benefit and caused the code to be more cluttered than it already is.
Added function to test input to make sure that the file exists and that the contents can be encrypted, i.e. it's not empty or it doesn't only contain numbers.